Fixed an issue with NULL json values
Whenever the bot is requested a search for a token with an unknown
market_cap_usd, as well as an unknown available_supply, the json value for those
specific keys would be null. The bot would then try to format a NoneType
with float() as a result and fail to search for the coin. It's fixed for
now, and I also applied minor comment changes.